Adds generic entity event generation (#184)
* Adds generic entity event generation

Adds the ability to generate generic entity events for the entity store.

This change introduces generic entity types and events, focusing on AWS resources.
It generates realistic resource names and ARNs based on the entity type,
and includes cloud metadata in the generated events.

* fix lint

* Adds .vscode to .gitignore

Excludes the .vscode directory from version control.

This prevents project-specific VS Code settings and workspace
files from being unintentionally committed to the repository.

* Reduces maximum random number for event offset

Reduces the maximum value of the randomly generated number used for the event date offset. This makes the random offset more realistic and prevents it from being excessively large.

* Adds event date offset functionality

Allows specifying an offset in hours for generated event timestamps.

This enables the generation of events that appear to have occurred
in the past, relative to the current time. A new parameter 'offsetHours'
is introduced to control the event date offset and defaults to 10 hours.

The 'getOffset' function now accepts an optional 'offsetHours' parameter,
providing flexibility in setting the offset.

* fix es lint errors